Back/donation/master -> back/master로 병합

See merge request s11-fintech-finance-sub1/S11P21A601!154

+++ backend/donation/src/main/java/kr/co/olivepay/donation/dto/req/DonationReq.java
@@ -16,7 +16,7 @@ public record DonationReq(
         String phoneNumber,
 
         @Positive(message = "금액은 양수인 정수여야 합니다.")
-        Integer money,
+        Long money,
 
         Long franchiseId,
